domicilios/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.2
djangorestframework==3.16.0
exceptiongroup==1.2.2
Faker==37.1.0
faker-vehicle==0.2.0
googlemaps==4.10.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
postgres==4.0
psycopg2-binary==2.9.10
psycopg2-pool==1.2
pytest==8.3.5
pytest-django==4.11.1
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
style==1.1.0
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
update==0.0.1
urllib3==2.4.0
